requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
